extracted createColumnsForSectionValue to seperate file
